requirements.txt
tqdm
numpy
scipy
pandas
psutil
pytest
torch==1.8.1
transformers==4.6.1
torch_optimizer==0.1.0
sympy==1.8
ray[default]==1.3.0
ray[tune]==1.3.0
PyYAML~=5.4.1
pycocoevalcap~=1.2
pycocotools~=2.0.2
regex~=2021.4.4
zss==1.2.0
git+https://github.com/google-research/bleurt.git
jupyter
bertviz
aioredis==1.3.1
